feat: add code for Fee FeeType FeeCategory and fix pom, skip test

+++ src/backend/resident_fee-service/src/main/java/com/project/entity/Fee.java
@@ -19,7 +19,7 @@ public class Fee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "FeeID")
-    private int feeId;
+    private Long feeId;
 
     /**
      * FeeType relationship.
